+++ src/modules/github/components/pull-list-item-status/types.ts
@@ -1,0 +1,5 @@
+export enum ECheckStatus {
+  FAILED = 'FAILED',
+  RUNNING = 'RUNNING',
+  SUCCESS = 'SUCCESS'
+}
