+++ src/modules/github/components/pull-list-item/pull-list-item.tsx
@@ -1,33 +1,10 @@
-import {
-  CheckIcon,
-  Cross2Icon,
-  DotFilledIcon,
-  ExclamationTriangleIcon
-} from '@radix-ui/react-icons'
-import { useQuery } from '@tanstack/react-query'
 import Link from 'next/link'
 
 import { Checkbox } from '@/elements/root/checkbox/checkbox'
-import { Skeleton } from '@/elements/root/skeleton/skeleton'
-import { Tooltip } from '@/elements/root/tooltip/tooltip'
 
-import { githubService } from '@/services/root/github'
-
-import { queryKey } from '@/utils/root/index'
+import { PullListItemStatus } from '@/components/github/root/pull-list-item-status/pull-list-item-status'
 
 import type { TPull } from '@/types/github/root/index'
-
-enum ECheckStatus {
-  FAILED = 'FAILED',
-  RUNNING = 'RUNNING',
-  SUCCESS = 'SUCCESS'
-}
-
-interface IPullListItemStatusProps {
-  isLoading: boolean
-  checkStatus: ECheckStatus | undefined
-  error: Error | null
-}
 
 interface IPullListItemProps {
   installationId: number
@@ -37,34 +14,6 @@ interface IPullListItemProps {
   handlePullCheckChange: (pullNumber: number) => void
 }
 
-export const PullListItemStatus = ({
-  isLoading,
-  checkStatus,
-  error
-}: IPullListItemStatusProps) => {
-  if (isLoading) {
-    return <Skeleton variant="rect" className="!h-full" />
-  }
-
-  if (error) {
-    return (
-      <Tooltip tooltip={error.message}>
-        <ExclamationTriangleIcon className="w-full h-full text-error" />
-      </Tooltip>
-    )
-  }
-
-  if (checkStatus === ECheckStatus.SUCCESS) {
-    return <CheckIcon className="w-full h-full text-success" />
-  }
-
-  if (checkStatus === ECheckStatus.RUNNING) {
-    return <DotFilledIcon className="w-full h-full text-info" />
-  }
-
-  return <Cross2Icon className="w-full h-full text-error" />
-}
-
 export const PullListItem = ({
   installationId,
   owner,
@@ -72,71 +21,6 @@ export const PullListItem = ({
   pull,
   handlePullCheckChange
 }: IPullListItemProps) => {
-  const {
-    data: checkStatus,
-    isLoading,
-    error
-  } = useQuery<ECheckStatus>({
-    queryKey: queryKey.github.pullStatus(owner, repo, pull.number),
-    queryFn: async ({ signal }) => {
-      return githubService
-        .listCommits(owner, repo, {
-          params: {
-            sha: pull.headRef,
-            perPage: 1
-          },
-          signal
-        })
-        .then(([latestCommit]) =>
-          Promise.all([
-            githubService
-              .listBranchRequiredStatusChecks(
-                owner,
-                repo,
-                pull.baseRef,
-                installationId
-              )
-              .then(({ contexts: requiredChecksName }) => requiredChecksName),
-            githubService
-              .listCheckRunsForRef(owner, repo, latestCommit!.sha ?? '')
-              .then(
-                ({ check_runs: latestCommitCheckRuns }) => latestCommitCheckRuns
-              )
-          ])
-        )
-        .then(([requiredChecksName, latestCommitCheckRuns]) => {
-          const hasRunningChecks = latestCommitCheckRuns.some(
-            (checkRun) =>
-              checkRun.status === 'queued' || checkRun.status === 'in_progress'
-          )
-
-          if (hasRunningChecks) {
-            return ECheckStatus.RUNNING
-          }
-
-          const hasNonSuccessChecks = latestCommitCheckRuns.some(
-            (checkRun) => checkRun.conclusion !== 'success'
-          )
-
-          if (hasNonSuccessChecks) {
-            return ECheckStatus.FAILED
-          }
-
-          const hasAllRequiredChecks = requiredChecksName.every((checkName) =>
-            latestCommitCheckRuns.some(
-              (checkRun) => checkRun.name === checkName
-            )
-          )
-
-          if (!hasAllRequiredChecks) {
-            return ECheckStatus.RUNNING
-          }
-
-          return ECheckStatus.SUCCESS
-        })
-    }
-  })
-
   return (
     <div className="flex gap-2 items-center">
       <Checkbox
@@ -156,9 +40,10 @@ export const PullListItem = ({
       </Link>
       <div className="w-6 h-6 flex items-center flex-none">
         <PullListItemStatus
-          isLoading={isLoading}
-          checkStatus={checkStatus}
-          error={error}
+          installationId={installationId}
+          owner={owner}
+          repo={repo}
+          pull={pull}
         />
       </div>
     </div>
